\name{as_igraph}
\alias{as_igraph}
\title{Convert osmar object to igraph}
\usage{
  as_igraph(obj)
}
\arguments{
  \item{obj}{An \code{\link{osmar}} object}
}
\value{
  An \link[igraph]{igraph-package} \code{graph} object
}
\description{
  Convert an osmar object to an igraph (see
  \link[igraph]{igraph-package}).
}

