\name{as_igraph0}
\alias{as_igraph0}
\title{Convert osmar object to igraph0}
\usage{
  as_igraph0(obj)
}
\arguments{
  \item{obj}{An \code{\link{osmar}} object}
}
\value{
  An igraph0-package \code{graph} object
}
\description{
  Convert an osmar object to an igraph0 (see
  igraph0-package).
}

