\name{osmar-package}
\alias{osmar-package}
\alias{osmar}
\docType{package}
\title{
  Importing and working with OpenStreetMapData
}
\description{
  This package imports the (Spatial and Non-Spatial) Data of OpenStreetMap into objects of R by using the OSM-API v0.6.
  For now it is reduced on using various commands of type "GET", like getting data of a explicit element or a region defined by coordinates.
  It also includes some methods on working with the newly created objects.
}
\details{
\tabular{ll}{
    Package: \tab osmar\cr
    Type: \tab Package\cr
    Version: \tab 0.13-4\cr
    Date: \tab 2011-07-06\cr
    License: \tab GPL-2\cr
    LazyLoad: \tab yes\cr
}

}
\author{
  Thomas Schlesinger
  
  Maintainer: Thomas Schlesinger <tho.schlesinger@googlemail.com>
  ~~ The author and/or maintainer of the package ~~
}
\references{
%~~ Literature or other references for background information ~~
}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
%~~ Optional links to other man pages, e.g. ~~
  \code{\link{getOSMObject}}
    \code{\link{XML}}
    \code{\link{RCurl}}
    \code{\link{sp}}
}
\examples{
%~~ simple examples of the most important functions ~~
}
